address Till's feedback
