+++ flink-core/src/main/java/org/apache/flink/configuration/JobManagerOptions.java
@@ -364,6 +364,17 @@ public class JobManagerOptions {
         Declarative
     }
 
+    @Documentation.Section({
+        Documentation.Sections.EXPERT_SCHEDULING,
+        Documentation.Sections.ALL_JOB_MANAGER
+    })
+    public static final ConfigOption<Integer> MIN_PARALLELISM_INCREASE =
+            key("jobmanager.declarative-scheduler.min-parallelism-increase")
+                    .intType()
+                    .defaultValue(0)
+                    .withDescription(
+                            "Configure the minimum increase in parallelism for a job to scale up.");
+
     /** Config parameter determining the declarative scheduler implementation. */
     @Documentation.ExcludeFromDocumentation("Declarative scheduler is still in development.")
     public static final ConfigOption<DeclarativeSchedulerType> DECLARATIVE_SCHEDULER_TYPE =
